requirements.txt
python-telegram-bot[aio]==20.4
aiohttp==3.9.3
